requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.9
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.21
lxml==5.3.1
marshmallow==3.26.1
multidict==6.3.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
